Fix jQuery conflict in WP plugins
So I was injecting my CSS/JS into the header/footer, which (as I
learned) circumvents the Wordpress system of queueing scripts/styles.
So I conformed to the way WP does it.

This was to fix a JS error on the events page when navigating to
"Previous Events".  It loads via ajax, and due to conflicting jQuery
versions, things didn't work right.  This changes fixes that issue, but
now going back to "Next Events" :(